arrowconv: move RowBytes marshal-match test into convert_test.go
